refactor: change the directory structure
