+++ src/features/groups/components/ContainerCardList/ContainerCard/ContainerCardDropdownMenuTriggerButton/DeleteContainerDialogContent.tsx
